+++ packages/std/src/math/conversion.rs
@@ -60,6 +60,267 @@ pub fn shrink_be_int<const INPUT_SIZE: usize, const OUTPUT_SIZE: usize>(
     Some(output)
 }
 
+/// Helper macro to implement `TryFrom` for a type that is just a wrapper around another type.
+/// This can be used for all our integer conversions where `bnum` implements `TryFrom`.
+macro_rules! forward_try_from {
+    ($input: ty, $output: ty) => {
+        impl TryFrom<$input> for $output {
+            type Error = $crate::ConversionOverflowError;
+
+            fn try_from(value: $input) -> Result<Self, Self::Error> {
+                value
+                    .0
+                    .try_into()
+                    .map(Self)
+                    .map_err(|_| Self::Error::new(stringify!($input), stringify!($output), value))
+            }
+        }
+    };
+}
+pub(crate) use forward_try_from;
+
+/// Helper macro to implement `TryFrom` for a conversion from a bigger signed int to a smaller one.
+/// This is needed because `bnum` does not implement `TryFrom` for those conversions
+/// because of limitations of const generics.
+macro_rules! try_from_int_to_int {
+    ($input: ty, $output: ty) => {
+        // statically assert that the input is bigger than the output
+        static_assertions::const_assert!(
+            core::mem::size_of::<$input>() > core::mem::size_of::<$output>()
+        );
+        impl TryFrom<$input> for $output {
+            type Error = $crate::ConversionOverflowError;
+
+            fn try_from(value: $input) -> Result<Self, Self::Error> {
+                $crate::math::conversion::shrink_be_int(value.to_be_bytes())
+                    .ok_or_else(|| Self::Error::new(stringify!($input), stringify!($output), value))
+                    .map(Self::from_be_bytes)
+            }
+        }
+    };
+}
+pub(crate) use try_from_int_to_int;
+
+/// Helper macro to implement `TryFrom` for a conversion from a unsigned int to a smaller or
+/// equal sized signed int.
+/// This is needed because `bnum` does not implement `TryFrom` for all of those conversions.
+macro_rules! try_from_uint_to_int {
+    ($input: ty, $output: ty) => {
+        // statically assert that...
+        // input is unsigned
+        static_assertions::const_assert_eq!(stringify!($input).as_bytes()[0], b'U');
+        // output is signed
+        static_assertions::const_assert_eq!(stringify!($output).as_bytes()[0], b'I');
+        // input is bigger than output (otherwise we would not need a `TryFrom` impl)
+        static_assertions::const_assert!(
+            core::mem::size_of::<$input>() >= core::mem::size_of::<$output>()
+        );
+
+        impl TryFrom<$input> for $output {
+            type Error = $crate::ConversionOverflowError;
+
+            fn try_from(value: $input) -> Result<Self, Self::Error> {
+                use bnum::prelude::As;
+                // $input::MAX has to be bigger than $output::MAX, so we can just cast it
+                if value.0 > Self::MAX.0.as_() {
+                    return Err(Self::Error::new(
+                        stringify!($input),
+                        stringify!($output),
+                        value,
+                    ));
+                }
+
+                // at this point we know it fits
+                Ok(Self(value.0.as_()))
+            }
+        }
+    };
+}
+pub(crate) use try_from_uint_to_int;
+
+#[cfg(test)]
+pub(crate) fn test_try_from_uint_to_int<I, O>(input_type: &'static str, output_type: &'static str)
+where
+    I: super::num_consts::NumConsts
+        + From<u32>
+        + Copy
+        + TryFrom<O, Error = crate::ConversionOverflowError>
+        + core::fmt::Debug
+        + core::ops::Add<Output = I>,
+    O: TryFrom<I, Error = crate::ConversionOverflowError>
+        + From<u32>
+        + super::num_consts::NumConsts
+        + core::cmp::PartialEq
+        + core::fmt::Debug,
+    String: From<I>,
+{
+    let v = I::MAX;
+    assert_eq!(
+        O::try_from(v),
+        Err(crate::ConversionOverflowError::new(
+            input_type,
+            output_type,
+            v
+        )),
+        "input::MAX value should not fit"
+    );
+
+    let max = I::try_from(O::MAX).unwrap();
+    assert_eq!(O::try_from(max), Ok(O::MAX), "output::MAX value should fit");
+
+    // but $output::MAX + 1 should not fit
+    let v = max + I::ONE;
+    assert_eq!(
+        O::try_from(v),
+        Err(crate::ConversionOverflowError::new(
+            input_type,
+            output_type,
+            v
+        )),
+        "output::MAX + 1 should not fit"
+    );
+
+    // zero should work
+    let v = I::ZERO;
+    assert_eq!(O::try_from(v), Ok(O::ZERO), "zero should fit");
+
+    // 42 should work
+    assert_eq!(
+        O::try_from(I::from(42u32)),
+        Ok(O::from(42u32)),
+        "42 should fit"
+    )
+}
+
+#[cfg(test)]
+pub(crate) fn test_try_from_int_to_uint<I, O>(input_type: &'static str, output_type: &'static str)
+where
+    I: super::num_consts::NumConsts
+        + From<i32>
+        + Copy
+        + TryFrom<O>
+        + core::fmt::Debug
+        + core::ops::Add<Output = I>,
+    O: TryFrom<I, Error = crate::ConversionOverflowError>
+        + From<u32>
+        + super::num_consts::NumConsts
+        + core::cmp::PartialEq
+        + core::fmt::Debug,
+    String: From<I>,
+    <I as std::convert::TryFrom<O>>::Error: std::fmt::Debug,
+{
+    if core::mem::size_of::<I>() <= core::mem::size_of::<O>() {
+        // if the input type is smaller than the output type, then `I::MAX` should fit into `O`
+        let v = I::MAX;
+        assert_eq!(
+            O::try_from(v),
+            Ok(O::try_from(v).unwrap()),
+            "input::MAX value should fit"
+        );
+    } else {
+        // if the input is bigger than the output, then `I::MAX` should not fit into `O`
+        let v = I::MAX;
+        assert_eq!(
+            O::try_from(v),
+            Err(crate::ConversionOverflowError::new(
+                input_type,
+                output_type,
+                v
+            )),
+            "input::MAX value should not fit"
+        );
+        // but `O::MAX` should fit
+        let max = I::try_from(O::MAX).unwrap();
+        assert_eq!(
+            O::try_from(max),
+            Ok(O::try_from(max).unwrap()),
+            "output::MAX value should fit"
+        );
+        // while `O::MAX + 1` should not
+        let v = max + I::ONE;
+        assert_eq!(
+            O::try_from(v),
+            Err(crate::ConversionOverflowError::new(
+                input_type,
+                output_type,
+                v
+            )),
+            "output::MAX + 1 should not fit"
+        );
+    }
+
+    // negative numbers should fail
+    let v = I::from(-42i32);
+    assert_eq!(
+        O::try_from(v),
+        Err(crate::ConversionOverflowError::new(
+            input_type,
+            output_type,
+            v
+        )),
+        "negative numbers should not fit"
+    );
+
+    // zero should work
+    let v = I::ZERO;
+    assert_eq!(O::try_from(v), Ok(O::ZERO), "zero should fit");
+
+    // 42 should work
+    assert_eq!(
+        O::try_from(I::from(42i32)),
+        Ok(O::from(42u32)),
+        "42 should fit"
+    )
+}
+
+/// Helper macro to implement `TryFrom` for a conversion from a signed int to an unsigned int.
+/// This is needed because `bnum` does not implement `TryFrom` for all of those conversions.
+macro_rules! try_from_int_to_uint {
+    ($input: ty, $output: ty) => {
+        // statically assert that...
+        // input is signed
+        static_assertions::const_assert_eq!(stringify!($input).as_bytes()[0], b'I');
+        // output is unsigned
+        static_assertions::const_assert_eq!(stringify!($output).as_bytes()[0], b'U');
+
+        impl TryFrom<$input> for $output {
+            type Error = ConversionOverflowError;
+
+            fn try_from(value: $input) -> Result<Self, Self::Error> {
+                use bnum::prelude::As;
+                // if $input::MAX is smaller than $output::MAX, we only need to check the sign
+                if core::mem::size_of::<$input>() <= core::mem::size_of::<$output>() {
+                    if value.is_negative() {
+                        return Err(ConversionOverflowError::new(
+                            stringify!($input),
+                            stringify!($output),
+                            value,
+                        ));
+                    }
+
+                    // otherwise we can just cast it
+                    Ok(Self(value.0.as_()))
+                } else {
+                    // $output::MAX is smaller than $input::MAX.
+                    // If it is negative or too big, we error.
+                    // We can safely cast $output::MAX to $input size
+                    if value.is_negative() || value.0 > <$output>::MAX.0.as_() {
+                        return Err(ConversionOverflowError::new(
+                            stringify!($input),
+                            stringify!($output),
+                            value,
+                        ));
+                    }
+
+                    // at this point we know it fits
+                    Ok(Self(value.0.as_()))
+                }
+            }
+        }
+    };
+}
+pub(crate) use try_from_int_to_uint;
+
 #[cfg(test)]
 mod tests {
     use super::*;